\name{bins}
\alias{bins}
\docType{data}

\title{
  bins data for GOVS
}

\description{
  bins data of 3515 bins for GOVS.
}

\references{
Liu H J, Wang X, Xiao Y, et al. CUBIC: an atlas of genetic architecture promises directed maize improvement[J]. Genome biology, 2020, 21(1): 1-17.
}

\usage{
  data(bins)
}