\name{binsInfo}
\alias{binsInfo}
\docType{data}



\title{
  bins information data for GOVS
}


\description{
  bins information data of 3515 bins for GOVS.
}


\usage{
  data(binsInfo)
}

\references{
Liu H J, Wang X, Xiao Y, et al. CUBIC: an atlas of genetic architecture promises directed maize improvement[J]. Genome biology, 2020, 21(1): 1-17.
}