\name{IBDTestData}
\alias{IBDTestData}
\docType{data}
\title{
Example data for IBD map construction
}
\description{
Example data for IBD map construction
}
\usage{data("IBDTestData")}
\format{
A list:
  \describe{
    \item{\code{snpParents}}{a matrix regarding genotype of parents.}
    \item{\code{markerInfo}}{a matrix or dataframe with four cols regarding information of markers.}
    \item{\code{snpProgeny}}{a character vector involes genotypes of progeny.}
    \item{\code{posGenetic}}{a numeric vector regarding genetic positions.}
  }
}
\details{
Details see: \cr
Liu H J, Wang X, Xiao Y, et al. CUBIC: an atlas of genetic architecture promises directed maize improvement[J]. Genome biology, 2020, 21(1): 1-17. \cr

https://github.com/heroalone/HMM-IBD
}
\source{
Liu H J, Wang X, Xiao Y, et al. CUBIC: an atlas of genetic architecture promises directed maize improvement[J]. Genome biology, 2020, 21(1): 1-17. \cr

https://github.com/heroalone/HMM-IBD
}
\references{
Liu H J, Wang X, Xiao Y, et al. CUBIC: an atlas of genetic architecture promises directed maize improvement[J]. Genome biology, 2020, 21(1): 1-17. \cr

https://github.com/heroalone/HMM-IBD
}
\examples{
  data(IBDTestData)
}