\name{MZ}
\alias{MZ}
\docType{data}


\title{
  hapmap data (genotypic data) for GOVS
}


\description{
  hapmap data of Zheng58 F1 progeny (genotypic data) for GOVS.
}


\usage{
  data(MZ)
}

\references{
Xiao Y, Jiang S, Cheng Q, et al. The genetic mechanism of heterosis utilization in maize improvement[J]. Genome Biology, 2021, 22(1): 1-29.
}