\name{phe}
\alias{phe}
\docType{data}

\title{
  Phenotypic data for GOVS
}


\description{
  Phenotypic data of Zheng58 F1 progeny for GOVS.
}


\usage{
  data(phe)
}

\references{
Xiao Y, Jiang S, Cheng Q, et al. The genetic mechanism of heterosis utilization in maize improvement[J]. Genome Biology, 2021, 22(1): 1-29.
}